Moved API functionality to a React context.
Added an API context that can be shared between components to allow them to communicate with the backend.
This centralises the communication with the backed and avoids having to have multiple files that depend on BASE_URL (nice to have this cleaned up).
In the future this will be a place where authentication with the backend will exist.