Removed the cloning of the gallery items when setting the computed thumb resolution.
This cloning isn't needed anymore because the gallery item is now designed to have the thumb resolution updated.
The cloning was creating an extra copy which was breaking the update of the description because the clone was being discarded when the gallery is rendered again later, thus discarding the update to the description.
The cloning would also be expensive, expecially when a large gallery is being viewed.