The user can now restart failed uploads
I was getting problems in the midst of uploading 3000+ assets.
I'm not sure what the problem is from the error message, but I'm hoping I can get through it by allowing failed uploads to be retried.
So a separate list is maintained with uploads that have failed. When the user clicks the retry button they are added back into the queue.
I've also added slightly nicer styling to the upload queue and status.