Fixed an issue uploading assets with unusual characters in the metadata.
I found an error uploading assets that have strange characters in the metadata that is passed to the backend through a header:

TypeError: Failed to execute 'setRequestHeader' on 'XMLHttpRequest': String contains non ISO-8859-1 code point.
    at xhr.js:164:19
    at Object.y [as forEach] (utils.js:276:12)
    at xhr.js:158:7
    at new Promise (<anonymous>)
    at a.register.e.exports (xhr.js:16:10)
    at a.register.e.exports (dispatchRequest.js:56:35)
    at a.register.h.request (Axios.js:109:15)
    at h.post (Axios.js:144:19)
    at Function.post (bind.js:9:15)
    at Ye.uploadAsset (api.ts:202:38)

The metadata in particular was a reverse geocoded location in a European country. This appears to be a browser limitation (at least in Chrome, untested in other browser) and I was unable to find a simple fix it for it.

I decided to restructure the upload process to separate the metadata into a separate HTTP POST request so the metadata can be passed to the backend via the body of the request instead of through a header. This fixes the problem and seems like a slightly better structure for asset upload than before.

A better way to do this would be a a multipart form upload that would combine the 4 request per asset into a single HTTP request, except the code would be much more complicated and there is value in keeping this simple for the moment. I might convert to a multipart form upload in the future.