Converted the test added in the previous commit to a proper test.
It has a better name and more specific expectations.
I let the code tell me what the right output is.
This seems kind of like cheating on TDD, but it's a very effective technique for creating tests for existing code when you believe that your code already works (and therefore you can trust the output).