Now sharing the user-interface between the web-based and Electron-based frontends.
I extracted a user-interface package to for the UI code.

I had to make changes to the exif library to get it compiling through
TypeScript.

For some reason, after converting to pnpm Parcel wants to put its
cache and dev deps at the root of the monorepo.

I had to update Tailwind configuration so that it can figure out the
styles that are used in the shared user-interface package.