Now uploading all files that are found on the local drive.
This is not the final behaviour. A user should be able to choose
which local files are uploaded. But this is a great stepping stone
that shows that I can use Node.js code in the Electron renderer process
to fine image files and then use my existing upload code (which relies
on the File object) to upload each image that is found.

I did a small refactor of the upload code to make it useable in the
new way.