Added a stub for the "Computer" page.
I'm not sure how to have a "Computer" that only exists in the Electron
and mobile versions.

I feel like I need to refactor the Main component so that the computer
page is injected, but I'm sure what is the simplest way to achieve this
and I could easily do a refactor that massively over complicates it.

Instead I'll hardcode a placeholder for the computer page directly into
the Main component then after that I'll decide the simplest way
to pull it up and out of the Main component so that it is only included
in the Electron and mobile versions.