Refactor to pull the Computer page up to the Electron package
This is the simplest possible refactor that pulls the Computer page up
so that it can be included in the Electron version, but not in the
web-based version (it doesn't make sense there).

I could have created a much more general abstraction here, but I'm not
sure how well it would work and I'm not sure that the benefit is
worth the effort.