Refactor to seperate "gallery context" from "gallery source".
I wanted to make the gallery work better with assets loaded from both
the cloud and the local computer.

So assets are now provided by a "gallery source". There's a separate
source for both cloud and local computer.

I pushed the code for "asset view" down into the gallery, which makes
the gallery easier to reuse.

I've removed pagination of assets and the infinite scroller. It was
going to be difficult to keep these working and not worth maintaining
then because in my storage database prototype I know that both of
those things will be completely different.

The search got moved to its own context because it has to sit above
the gallery context now.