Committed Android file scanning code
This code was imported from the prototype in the code repo
prototype-capacitor-file-scanning.

It is Java code to run on Android that scans the file system for images
and attempts to upload them.

There are changes to be made to this code but this a great starting
point that has already been tested independently.