Scanned files on iOS now displayed in gallery
Removed iOS uploading code because it isn't needed yet.

Now showing scanned files in the gallery in the Computer page.

I had to change the names of the functions checkPermissions
and requestPermissions because names seem to have special
meaning and I couldn't get the function calls working
with the change.

Also it doesn't seem to be able to call async functions.