Big refactor to introduce "source" and "sink".
This refactor is a prelude to using indexeddb to store assets and
metadata locally.

To enable a movement in this direction I've consolidated all the
gallery input (source) and gallery output (sink) into source and
sink abstractions. At the moment both the source and sink are
implemented by the "cloud source" but later I would like to have
separate cloud source and sink and separate "local" source and sink.

Chained together cloud and local source/sink should allow assets to
be read locally and failing that, read from the cloud. For updates
they will first be made locally and then via the cloud.

Replaced add/remove label and set description operations with a single
generic "patch metdata" operation.