Added "database operations".
Asset metadata is updated now trough a sequence of set/push/pull
operations. This is required so that a set of locally applied
operations can easily be sync'd to the cloud and then to other
devices.

Operations are inspired by MongoDB.

An asset is created now simply by updating its metadata the first time.