Moves the assets state into gallery context.
Refactor to move responsibility of maintaining asset state into
the gallery context.

It makes more sense to do this here than in the source, reducing
the responsibilty of the source to just retrieving data rather than
doing that and maintaining state.

This change make things fit together more elegantly.

Gallery items are mow passed all the way down to gallery layout
via the context.

The soure and sink are encapsulated within the gallery context.