Converted cloud gallery source/sink to a custom hook.
Converted cloud gallery source/sink from context to just a normal
React custom hook. There was no need for these to remain as context.

This had to be changed now because the upload context depends on the
gallery context, so the gallery context had to be moved up to the
app level. This makes sense because the gallery context now manages
the displayed list of assets and I want this to be persistent across
pages in the application.

With the gallery context moving up to the app level I had to find a
different way to implement the cloud source/sink contexts, but it
works out better for them to be a custom hook because they don't
(currently) need to be accessed from anywhere in the UI hierarchy.

I had to restruct the setup in the app component so that the
new hooks for source/sink could depend on the api context.