Indexeddb experimental update.
Experimental version that connects the gallery only to indexeddb.

I was finding it difficult to figure out how to get indexedb with
fallback to the cloud, especially the part when metadata needs to
be syncrhonized from cloud to load and vice versa.

It seems like a simple step and to help build understanding to
just get the gallery working with indexeddb and then figure out
how to get the cloud fallback and sync working.