Implementing local storage with cloud fallback.
Local uploads and updates are stored in a queue in indexeddb.

Periodically the queue is flushed and the uploads/updates are
propagated to the backend.

I've removed the `addAsset` function from the sink. It makes things
a lot easier if the `updateAsset` function can just upsert assets
that don't exist.

Simplified the way `loadAsset` works to not use a callback. Making
it async-based instead makes it a bit cleaner to use.