Implementation of two way database synchronization.
Journal records in the backend are stored by their timestamp
that is counting down to the year 3000. This is required so that
the most recent updates are loaded first.

Added database fixtures for journal entries.

Collection id is now passed all the way through sinks and the api.

I've restructured the indexeddb so that it can store multiple
collections. I have to use a separate database for each collection.