Initial sync when the database is empty
When the database is empty (specifically when the metadata collection
for a database has no records) the entire set of asset metadata
is now downloaded into local storage on startup.

For each collection that is retrieved, the latest update id is stored
to prevent updates being read again. There is no need for it when we
getting the entire latest snapshot of data.