Updated to the service-worker to what I know works.
I prototyped this code in a seperate minimal web page so I know that
it works well for my use cases.

It has a network-first caching policy. It resorts to cached versions of
core assets when the network is not available.

I converted from TypeScript to JavaScript for the service worker
because it's just too hard to get the typing right.

I'm not packing the JavaScript file for the service worker anymore,
just copying it over directly to the Webpack output folder.

The bundle name now includes a cache busting content hash.