Big restructure of asset operations
Asset and operations are now structure in a more flat fashion as
opposed to nested.

Operations from different client might be interleaved, so it makes
sense to store collectionId and assetId against every journal record
that is retrieved from the backend.

When a client asks for updates, those updates originally from itself
are now filtered out in the backend rather than being returned
to the client and filtered out there.