Factored out a persistent queue.
This allows removal of getLeastRecentRecord from the
IDatabaseCollection interface which is good because it's unecessary
to support this under AWS S3 storage.

Fixed the id generator for the queue so that records can be retrieved
in a FIFO manner. The id is a reverse chronological timestamp.

I had to add a rotating decrementing sequence number to ensure
timestamps are unique.