Moved to checkAsset to source.
Moved to checkAsset functions to become checkAssets in the gallery
source. These functions now deal correclty with hash collisions.

These functions didn't belong in the sink, because they are about
writing data (what the source does) and not about reading data
(which is what the sink does).

It should be extremely rare for a hash collision to occur, but if two
different assets (or more) are logged against the same hash that hash
now lists each of them.