Restructure to generalize and test the sync code.
Renamed gallery source and sink to asset source and sink. These are
more about the "asset database" now than they are about the gallery!

Moved the database sync code to the database library and set it up
to be able to have automated testing.

I've removed getAssets and submitOperations from the source and sink
interfaces. It just seems to be duplicating functionality that
the database interface can already deal with.

I've added a cloud version of the database that uses the api to
interact with the database.