Getting and updating asset metadata moved back to source/sink.
This is a bit of a reversal of gears. But I realised that the
gallery context shouldn't directly deal with the database. It's much
better if the gallery is kept at a distance from them database so
I've moved metadata loading and updates back into source/sink.

So the database refactor was good and the database abstraction is
useful, but I went a bit too far trying to connect the gallery context
directly to the database. It simply feels better.