Now cloning sorted assets.
This has to be done to avoid sorting in place on the list of assets
currently be loaded during file scanning.

It's also required that setting state needs a new array. If I set the
same array (even when it's had new items added to it) it doesn't
trigger rendering because React thinks nothing has changed!