Big conversion back to MongoDB for the backend.
I've rolled back to the big conversion to a storage-based database.

I realised that there is a lot more work to do to make the storage-
based database performant and I just want to ship Photosphere.

There is now a minimal defs library for sharing types between
backend and frontend.

The database library has been collapsed back into user-interface
library.

I'm still using the database, source and sink abstractions in the
frontend, but I'm no longer trying to make the database abstract
work for the backend as well. The database is just too different
between backend and frontend to make the abstraction cover them both.