Restructure to a single indexeddb database.
I've collapsed the multiple indexeddb database into a single one
that separates assets by the setId field just like it works in the
MongoDB in the backend.

This makes the code in the frontend so much simpler. It also simplifies
the database operations that are passed to the backend.

I've added a fixture with assets in two separate collections.