Big restructure so that "gallery source" is a context with state.
I'm trying to solve the following problem:

I need to load all assets paginated so that they can be displayed
before they are fully loaded.

I need to support multiple sources: cloud+indexeddb and local files.

Even though I really liked the abstraction through the source/sink
interfaces the path of least resistance is to make the "source"
a React context so that loaded assets are state so that incremental
paginated loading triggers a render. Also so (in the future) updates
to the assets in memory also triggers a render.

At this points it makes sense to merge source and sink into the one
context, it just makes everything easier. I toyed with the idea of
keeping the database-sync in a separate context but it also ended up
making sense to integrate it into the one context.

Later I'll add a separate read-only "source" to support rendering
assets in the gallery from the local file system.

This really simplifies the setup of the App component.