Restructure so that assets are updated in memory at the top level.
The gallery is updated completely when any update to `assets`
happens: applying search, sort and layout.

Some of this feels like it will lead to bad performance. For example,
why update the entire gallery when editing the description for one
asset? I've decided to do the simplest setup possible and only if
it performs badly find a way to make it fast. (E.g the layout or parts
of the layout could be cached).

This setup also paves the way for edits to assets to come in from the
networ and trigger a render.