Removed setIndex from gallery item.
It turns out this can be removed if the data structure for asset look
ups is a hash rather than an array. It doesn't actually need to be an
array.

Note that we still have the "searched assets" array that is used to
navigate back/forward through the list of displayed (searched) assets.