Now recording the ID of the selected asset instead of the asset itself.
This works better when extracting the assetId from the URL because
at the point where that happens the asset in question isn't loaded.

So we need to select the asset by it's ID and a moment later when it's
loaded into memory we can then open the asset view for it.