Now loading assets through indexeddb in the same way as images.
The Video UI component can't depend on setId (the asset database)
and still be able to render videos from the local file system.

So I've made videos work the same was images. If the video is already
in indexeddeb it is loaded from there. If not existing locally,
the video is downloaded and stored in indexeddb.

It has to be this way in order to have a video locally before it
has been uploaded to the cloud.

There will come up a time when videos are too big for local storage,
if this happens I'll have to break them up or down scale them so they
do fit.