Replaced flexsearch with my own simple search algorithm.
I need an incrementally loaded search index and to also be able
to apply search to increments as they come in. Flexsearch doesn't
really handle this and also produces searched assets in an unsorted
order.

Also I didn't like it with flexsearch how the results come back in
unsorted order and a sort must then be applied to it. With this change
the searched assets come back in sorted order so there is no need for
separate sorting pass.