More efficient implementation for incrementally loading assets.
Now using Rx.js style observables/subscriptions to notify the gallery
layout of new items that have been loaded, allowing the layout to be
incrementally built as new assets come in.

This is much faster than how it used to be. Previously the layout
was computed from the entire asset set on each increment and it
grows with each new page of assets loaded.

Eventually the asset set will equal the entire asset library and so as
pages are loaded the layout was getting more and more expensive
to build. Now the layout is computed at each increment only for the
new assets in the increment and not for the entire growing set.