Fixed a really tricky state issue.
After changing "group by" it seems like new items were coming in during incrementable loading and using the old "group by" that was  supposedly updated in state but maybe hasn't come through before the newItems event is triggered.

I've changed the particular state to a ref and am using an artificial state variable to trigger rerender. Not ideal, but generally this seems to be a safer way to triger a rerender (maybe even better peformance as well if all state gets simplified in this way).