The upload script is now writing directly to storage to avoid problems uploading very large files.
I could spend a lot of time on this, but the upload script is only a means to an end: Building my own asset database.

Eventually if the peer-to-peer version of Photosphere comes about this will look very different anyway. There will be no upload scirpt, we only have to build and update the local database and then sync it to whereever it needs to go.