Big restructure to remove the need for content-type to be stored along side each asset.
This eliminates the need for the .info stored in the file system. That's good because storing these small files (33 bytes each) in the Linux file system stores them as 4k since that's the min file size.  That was a large wastage of  storage space.

The downside of this is that we no longer have the "content type" per asset stored in file system replicas. This information is still stored in cloud storage databases. We still store the content type of each main asset in the database, but no longer store the content type for display assets and thumbnails. Both of these are currently hardcoded to jpeg so it doesn't matter so much. If we ever need to record the content type of display and thumbnail we might just have to add new fields for that to the database record for each asset.