cli: Better performance: When adding local files to the database don't copy the file to temporary file to validate it.
Refactors the file scanner to not depend on the storage abstraction. The abstraction isn't needed in this case (just resulting in a duplicated walkDirectory function) but makes it easier to indicate that a file is local and so doesn't need to be downloaded before validation. Previously every file was being copied to a temporary file (even when already local) before validating it. Now only files contained in zip files are copied to a temporary file for validation.