Write locking is now implementeord in the local file system.
Because its only processes on the same machine that will modify shared files in the database (local, cloud or network) that means all lock files can be held in the local file system (Photosphere temp directory).

This removes the need for lock files in cloud and network storage and simplifies the code.