Failed attempt to get actual bun workers to process the task queue.
But it's not possible to define task handlers in the main thread and have them accessible in the worker.

The task handlers would have to be registered in the worker thread and wouldn't have access outside the closure in the verify function.

This also means the verify function wouldn't be portable to the browser.